requirements.txt
Flask
flask-cors
torch
torchvision
Pillow
torchaudio 
tensorflow-cpu
tensorflow 